Extract createNavigator to utils
