+++ test/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './createCollection';
+export * from './createNavigator';
 export * from './createPlayground';
 export * from './createScopeApi';
 export * from './createSource';
